+++ src/Eng/Scene/Comp/Lightmap.cpp
@@ -7,6 +7,4 @@ void Lightmap::Read(const JsObjectP &js_in, Lightmap &lm) {
     lm.size[0] = lm.size[1] = int(js_res.val);
 }
 
-void Lightmap::Write(const Lightmap &lm, JsObjectP &js_out) {
-    js_out.Push("res", JsNumber{ (double)lm.size[0] });
-}
+void Lightmap::Write(const Lightmap &lm, JsObjectP &js_out) { js_out.Push("res", JsNumber{lm.size[0]}); }
