+++ src/Eng/Scene/Comp/VegState.cpp
@@ -30,18 +30,18 @@ void VegState::Read(const JsObjectP &js_in, VegState &vs) {
 
 void VegState::Write(const VegState &vs, JsObjectP &js_out) {
     if (vs.movement_scale != 1.0f) {
-        js_out["movement_scale"] = JsNumber{double(vs.movement_scale)};
+        js_out["movement_scale"] = JsNumber{vs.movement_scale};
     }
 
     if (vs.tree_mode != 1.0f) {
-        js_out["tree_mode"] = JsNumber{double(vs.tree_mode)};
+        js_out["tree_mode"] = JsNumber{vs.tree_mode};
     }
 
     if (vs.bend_scale != 1.0f) {
-        js_out["bend_scale"] = JsNumber{double(vs.bend_scale)};
+        js_out["bend_scale"] = JsNumber{vs.bend_scale};
     }
 
     if (vs.stretch != 0.0f) {
-        js_out["stretch"] = JsNumber{double(vs.stretch)};
+        js_out["stretch"] = JsNumber{vs.stretch};
     }
 }
